requirements.txt
FastAPI
pydantic
sqlalchemy==2.0.18
aiosqlite
django
django-rest-framework